backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
groq==0.8.0
langchain==0.1.16
langchain-groq==0.1.3
python-multipart==0.0.9
